src/main/java/subway/service/domain/Station.java
package subway.service.domain;

import java.util.Objects;

public class Station {

    private long id;
    private final String name;

    public Station(final String name) {
        this.name = name;
    }

    public Station(final long id, final String name) {
        this.id = id;
        this.name = name;
    }

    public long getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public boolean hasSameName(Station station) {
        return this.name
                .equals(station.name);
    }

    @Override
    public boolean equals(final Object o) {
        if (this == o) return true;
        if (!(o instanceof Station)) return false;
        Station station = (Station) o;
        return id == station.id && Objects.equals(name, station.name);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, name);
    }

    @Override
    public String toString() {
        return "Station{" +
                "id=" + id +
                ", name='" + name + '\'' +
                '}';
    }

}
